+++ docs/src/changelog.md
@@ -67,6 +67,7 @@ project adheres to [Semantic Versioning](https://semver.org/).
     - Feed subscription edition form
 - Homogenize edition form layout, use horizontal fields
 - Rework deletion forms as modals
+- Map domain errors to user-facing error messages
 
 ## [v0.6.1](https://github.com/virtualtam/sparklemuffin/releases/tag/v0.6.1) - 2026-04-23
 ### Security